Extract zoom into an extension (Android)
