+++ package/android/src/main/java/com/mrousavy/camera/core/CameraSession.kt
@@ -30,7 +30,7 @@ import com.mrousavy.camera.extensions.capture
 import com.mrousavy.camera.extensions.createCaptureSession
 import com.mrousavy.camera.extensions.createPhotoCaptureRequest
 import com.mrousavy.camera.extensions.openCamera
-import com.mrousavy.camera.extensions.zoomed
+import com.mrousavy.camera.extensions.setZoom
 import com.mrousavy.camera.frameprocessor.FrameProcessor
 import com.mrousavy.camera.parsers.Flash
 import com.mrousavy.camera.parsers.Orientation
@@ -480,20 +480,7 @@ class CameraSession(
 
     // Zoom
     val cameraCharacteristics = cameraManager.getCameraCharacteristics(cameraId!!)
-    val zoomRange = (
-      if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) {
-        cameraCharacteristics.get(CameraCharacteristics.CONTROL_ZOOM_RATIO_RANGE)
-      } else {
-        null
-      }
-    ) ?: Range(1f, cameraCharacteristics.get(CameraCharacteristics.SCALER_AVAILABLE_MAX_DIGITAL_ZOOM) ?: 1f)
-    val zoomClamped = zoomRange.clamp(zoom)
-    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) {
-      captureRequest.set(CaptureRequest.CONTROL_ZOOM_RATIO, zoomClamped)
-    } else {
-      val size = cameraCharacteristics.get(CameraCharacteristics.SENSOR_INFO_ACTIVE_ARRAY_SIZE)!!
-      captureRequest.set(CaptureRequest.SCALER_CROP_REGION, size.zoomed(zoomClamped))
-    }
+    captureRequest.setZoom(zoom, cameraCharacteristics)
 
     // Torch Mode
     val torchMode = if (torch == true) CaptureRequest.FLASH_MODE_TORCH else CaptureRequest.FLASH_MODE_OFF
