+++ package/android/src/main/java/com/mrousavy/camera/extensions/CameraDevice+createPhotoCaptureRequest.kt
@@ -4,8 +4,6 @@ import android.hardware.camera2.CameraCharacteristics
 import android.hardware.camera2.CameraDevice
 import android.hardware.camera2.CameraManager
 import android.hardware.camera2.CaptureRequest
-import android.os.Build
-import android.util.Range
 import android.view.Surface
 import com.mrousavy.camera.parsers.Flash
 import com.mrousavy.camera.parsers.Orientation
@@ -87,20 +85,7 @@ fun CameraDevice.createPhotoCaptureRequest(
     }
   }
 
-  val zoomRange = (
-    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) {
-      cameraCharacteristics.get(CameraCharacteristics.CONTROL_ZOOM_RATIO_RANGE)
-    } else {
-      null
-    }
-  ) ?: Range(1f, cameraCharacteristics.get(CameraCharacteristics.SCALER_AVAILABLE_MAX_DIGITAL_ZOOM) ?: 1f)
-  val zoomClamped = zoomRange.clamp(zoom)
-  if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) {
-    captureRequest.set(CaptureRequest.CONTROL_ZOOM_RATIO, zoomClamped)
-  } else {
-    val size = cameraCharacteristics.get(CameraCharacteristics.SENSOR_INFO_ACTIVE_ARRAY_SIZE)!!
-    captureRequest.set(CaptureRequest.SCALER_CROP_REGION, size.zoomed(zoomClamped))
-  }
+  captureRequest.setZoom(zoom, cameraCharacteristics)
 
   captureRequest.addTarget(surface)
